+++ service/coupon/coupon-api/src/main/java/radiata/service/coupon/api/controller/CouponController.java
@@ -93,68 +93,4 @@ public class CouponController {
             .body(success(DELETE_COUPON.getMessage()));
     }
 
-    @PostMapping("/coupons/{couponId}/issue")
-    public ResponseEntity<? extends CommonResponse> issueCoupon(
-        @PathVariable
-        String couponId,
-        @RequestHeader("X-UserId")
-        String userId
-    ) {
-
-        couponApiService.issueCoupon(couponId, userId);
-
-        return ResponseEntity.status(ISSUE_COUPON.getHttpStatus())
-            .body(success(ISSUE_COUPON.getMessage()));
-    }
-
-    @PatchMapping("/couponissues/{couponIssueId}")
-    public ResponseEntity<? extends CommonResponse> useIssueCoupon(
-        @PathVariable
-        String couponIssueId,
-        @RequestHeader("X-UserId")
-        String userId
-    ) {
-
-        return ResponseEntity.status(USE_COUPON_ISSUE.getHttpStatus())
-            .body(success(USE_COUPON_ISSUE.getMessage(), couponApiService.useCouponIssue(couponIssueId, userId)));
-    }
-
-    @GetMapping("/couponissues/{couponIssueId}")
-    public ResponseEntity<? extends CommonResponse> getCouponIssue(
-        @PathVariable
-        String couponIssueId,
-        @RequestHeader("X-UserId")
-        String userId
-    ) {
-
-        return ResponseEntity.status(GET_COUPON_ISSUE.getHttpStatus())
-            .body(success(GET_COUPON_ISSUE.getMessage(), couponApiService.getCouponIssue(couponIssueId, userId)));
-    }
-
-    @GetMapping("/couponissues")
-    public ResponseEntity<? extends CommonResponse> getCouponIssues(
-        @RequestHeader("X-UserId")
-        String userId,
-        @PageableDefault(size = 10, sort = "createdAt", direction = DESC)
-        Pageable pageable
-    ) {
-
-        return ResponseEntity.status(GET_COUPON_ISSUES.getHttpStatus())
-            .body(success(GET_COUPON_ISSUES.getMessage(), couponApiService.getCouponIssues(userId, pageable)));
-    }
-
-    @DeleteMapping("/couponissues/{couponIssueId}")
-    public ResponseEntity<? extends CommonResponse> deleteCouponIssue(
-        @PathVariable
-        String couponIssueId,
-        @RequestHeader("X-UserId")
-        String userId
-    ) {
-
-        couponApiService.deleteCouponIssue(couponIssueId, userId);
-
-        return ResponseEntity.status(DELETE_COUPON_ISSUES.getHttpStatus())
-            .body(success(DELETE_COUPON_ISSUES.getMessage()));
-    }
-
 }
